prompts/requirements.txt
# Prompt Pipeline Dependencies
# ============================

# LLM APIs
openai>=1.0.0
anthropic>=0.7.0

# Configuration
pyyaml>=6.0

# Terminal colors
colorama>=0.4.6

# Optional: for testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

